Basic handling of headings...
However i'm not really sure this is the way to go.
Its not really just lexing? i'm already parsing the heading.
I'll have to think on this one.
But maybe its fine if i can parse the basic stuff as I go.
Meanwhile for the more complex constructs I would leave that until a further stage.