Changed over to a wrapper around the iter...
This provides more conveniance as we dont need to call pos.increment manually.
Same with nth, this make it less error prone.
However this means that calls to iter.next() increment the index even when we hit None!
I should have corrected all the indexes to account for this.
Though its possible somethings slipped through.
There are some indexes that need to be double checked anyway.
Also removed remains of TokenType::Whitespace.