Changed default theme, codeblock changes...
Eliminate if stament when parsing the language for a codeblock.
The language class for html now has a space in it to allow two classes to style the element.
This allows me to have a base language class that does all common elements.
Which is used in conjunction with a "python" class for ex which sets the rest of the highlighting.